requirements/constraints.txt
# Version constraints for pip-installation.
#
# This file doesn't install any packages. It specifies version constraints
# that will be applied if a package is needed.
#
# When pinning something here, please provide an explanation of why.  Ideally,
# link to other information that will help people in the future to remove the
# pin when possible.  Writing an issue against the offending project and
# linking to it here is good.

# Already in python3 standard library
futures; python_version == "2.7"

# TODO: Add constraint explanation
pylint==2.5.0
pycodestyle==2.5.0

# Keep same platform version
celery==3.1.26.post2
djangorestframework==3.9.4
django-filter==2.2.0
django-model-utils==4.0.0
django-oauth-toolkit==1.3.2
django-waffle==0.18.0
edx-drf-extensions==6.0.0
edx-opaque-keys==2.1.0
edx-proctoring==2.4.0
factory-boy==2.10.0
jsonfield2==3.0.3

